refactor(core): make library root authoritative
